refactor: add state to useSearchContext hook
